refactor(backend): rename "clients" to "backend"
BREAKING CHANGE: renamed Client `.http` option to `.backend`
