+++ lib/backends/fetch.test.js
@@ -3,14 +3,14 @@ const { expect } = require('chai')
 const fetch = require('node-fetch')
 const nock = require('nock')
 
-const FetchClient = require('./fetch-client')
+const Fetch = require('./fetch')
 
 const url = 'https://foo.com'
 
 describe('lib.fetch-client', () => {
   describe('FetchClient', () => {
     it('GETs the expected value', done => {
-      const client = new FetchClient({ fetch, url })
+      const client = new Fetch({ fetch, url })
 
       nock(url)
         .get('/magic')
@@ -31,7 +31,7 @@ describe('lib.fetch-client', () => {
     })
 
     it('POSTs the expected value', done => {
-      const client = new FetchClient({ fetch, url })
+      const client = new Fetch({ fetch, url })
 
       nock(url)
         .post('/magic', {
