+++ lib/client.js
@@ -1,10 +1,10 @@
 const Component = require('./loader')
-const Request = require('./request')
+const Request = require('./backends/request')
 
 class Client {
   constructor (options) {
-    const http = options.http || new Request(options.config)
-    const root = new Component({ splits: [], http: http, getNames: options.getNames })
+    const backend = options.backend || new Request(options.config)
+    const root = new Component({ splits: [], backend, getNames: options.getNames })
     if (options.spec) root._addSpec(options.spec)
     return root
   }
