+++ lib/loader.js
@@ -65,8 +65,8 @@ class Component {
         // E.g., assume `/foo/{name}/{property}` cannot exist.
         //
         const namedComponent = new Component({
+          backend: component.backend,
           getNames: options.getNames,
-          http: component.http,
           splits: splits,
           parameters: Object.assign({ [options.template]: name }, options.parameters)
         })
@@ -93,7 +93,7 @@ class Component {
     component.parameters = options.parameters
     component.template = options.template
     component.splits = options.splits.slice()
-    component.http = options.http
+    component.backend = options.backend
     component.getNames = options.getNames || (split => [ split ])
     component.children = []
     return component
@@ -162,7 +162,7 @@ class Component {
       if (!(split in parent)) {
         const component = new Component({
           getNames: this.getNames,
-          http: this.http,
+          backend: this.backend,
           parameters: this.parameters,
           template: template,
           splits
@@ -221,7 +221,7 @@ class Component {
    */
   _requestAsync (method, options) {
     const operationId = this.operationIds[method.toLowerCase()]
-    return this.http.http(Object.assign({
+    return this.backend.http(Object.assign({
       method,
       operationId,
       pathname: this.getPath(),
